Write new nextQuestion logic
Now if we run out of questions, we fetch some more.

However, this is only a temporary solution. For one, we aren't handling rate limits. Also, there's no way for us to know if we are "out" of fresh questions.

In the future, we could set it up so that if we try to get new questions multiple times and get lots of collisions (as in, the new questions are ones we've already seen), then we indicate to the user that they've seen all of the questions in the currently selected categories.